requirements.txt
attrs==19.3.0
autopep8==1.5
beautifulsoup4==4.8.2
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
cymem==2.0.3
en-core-web-sm==2.2.5
idna==2.8
importlib-metadata==1.5.0
lxml==4.5.0
more-itertools==8.2.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.1
packaging==20.1
pdfminer.six==20200124
pdfquery==0.4.3
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
py==1.8.1
pycodestyle==2.5.0
pycryptodome==3.9.6
pyparsing==2.4.6
pyquery==1.4.1
requests==2.22.0
roman==3.2
rope==0.16.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
textblob==0.15.3
thinc==7.3.1
tqdm==4.42.1
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.8
zipp==2.1.0
